refactor: use dependencies directly

+++ internal/entities/entity.go
@@ -1,5 +1,5 @@
 package entities
 
 type Entity interface {
-	GetId() int64
+	GetId() uint
 }
